chore: move table utility methods from AbstractTableTab to AbstractTab
renderTableScrollbar and handleTableClick are used by both AbstractTableTab
subclasses and multi-table AbstractTab subclasses (BrowseTab, HistoryTab,
RoutesTab, etc). Moving them to the common ancestor avoids awkward
cross-class static calls.
