+++ dsl/camel-jbang/camel-jbang-plugin-tui/src/main/java/org/apache/camel/dsl/jbang/core/commands/tui/AbstractTableTab.java
@@ -24,7 +24,6 @@ import dev.tamboui.terminal.Frame;
 import dev.tamboui.text.Span;
 import dev.tamboui.tui.event.KeyEvent;
 import dev.tamboui.tui.event.MouseEvent;
-import dev.tamboui.widgets.scrollbar.Scrollbar;
 import dev.tamboui.widgets.scrollbar.ScrollbarState;
 import dev.tamboui.widgets.table.TableState;
 
@@ -115,46 +114,5 @@ abstract class AbstractTableTab extends AbstractTab {
 
     protected void renderScrollbar(Frame frame, int rowCount) {
         renderTableScrollbar(frame, lastTableArea, tableState, tableScrollState, rowCount);
-    }
-
-    static void renderTableScrollbar(
-            Frame frame, Rect tableArea, TableState tableState, ScrollbarState scrollState, int rowCount) {
-        if (tableArea == null || tableState == null || scrollState == null) {
-            return;
-        }
-        int visibleRows = tableArea.height() - 3;
-        if (visibleRows <= 0 || rowCount <= visibleRows) {
-            return;
-        }
-        Rect scrollRect = new Rect(
-                tableArea.x() + tableArea.width() - 1,
-                tableArea.y() + 2,
-                1,
-                visibleRows);
-        scrollState.contentLength(rowCount);
-        scrollState.viewportContentLength(visibleRows);
-        scrollState.position(tableState.offset());
-        frame.renderStatefulWidget(Scrollbar.builder().build(), scrollRect, scrollState);
-    }
-
-    static boolean handleTableClick(MouseEvent me, Rect tableArea, TableState tableState, int rowCount) {
-        if (tableArea == null || tableState == null || rowCount <= 0) {
-            return false;
-        }
-        if (!me.isClick()) {
-            return false;
-        }
-        int mx = me.x();
-        int my = me.y();
-        if (mx < tableArea.x() || mx >= tableArea.x() + tableArea.width()
-                || my < tableArea.y() || my >= tableArea.y() + tableArea.height()) {
-            return false;
-        }
-        int rowIndex = tableState.offset() + (my - tableArea.y() - 2);
-        if (rowIndex >= 0 && rowIndex < rowCount) {
-            tableState.select(rowIndex);
-            return true;
-        }
-        return false;
     }
 }
